requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Database
sqlite-utils>=3.0

# Scheduling
APScheduler>=3.10.4

# Telegram
python-telegram-bot>=20.0

# Logging
colorlog>=4.8.1

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Date/Time
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1